[ces] 예약하기, 예약취소 조건 추가

+++ semi_project/src/main/java/com/gn/reservation/controller/DeleteReservationServlet.java
@@ -8,9 +8,12 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
+import com.gn.plan.vo.Plan;
 import com.gn.reservation.service.ReservationService;
 import com.gn.reservation.vo.Reservation;
+import com.gn.user.vo.User;
 
 
 @WebServlet("/deleteReservation")
@@ -24,22 +27,36 @@ public class DeleteReservationServlet extends HttpServlet {
 
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		String temp = request.getParameter("reservation_no");
-		int reservationNo = 0;
-		if(temp != null)
-			reservationNo = Integer.parseInt(temp);
-		Reservation reservation = new Reservation();
-		reservation.setReservationNo(reservationNo);
-		
-		int result = new ReservationService().deleteReservation(reservation);
-		if(result > 0) {
-			// 성공시 페이지 안정했음
-			RequestDispatcher view = request.getRequestDispatcher("/views/reservation/deleteSuccess.jsp");
-			request.setAttribute("reservation", reservation);
-			view.forward(request, response);
+		HttpSession session = request.getSession(false);
+		// 세션확인
+		if(session != null && session.getAttribute("user") != null) {
+			User user = (User)session.getAttribute("user");
+			int userNo = user.getUserNo();
+			
+			String temp = request.getParameter("planNo");
+			int planNo = 0;
+			if(temp != null)
+				planNo = Integer.parseInt(temp);
+			
+			Reservation reservation = Reservation.builder()
+					.userNo(userNo)
+					.planNo(planNo)
+					.build();
+			
+			reservation = new ReservationService().selectReservationDetail(reservation);
+			
+			int result = new ReservationService().deleteReservation(reservation);
+			if(result > 0) {
+				// 성공시 페이지 안만들었음
+				RequestDispatcher view = request.getRequestDispatcher("/views/reservation/reservationList.jsp");
+				request.setAttribute("reservation", reservation);
+				view.forward(request, response);
+			} else {
+				// 실패시 페이지 만들어야함
+				request.getRequestDispatcher("/views/reservation/deleteFail.jsp");
+			}
 		} else {
-			// 실패시 페이지 만들어야함
-			request.getRequestDispatcher("/views/reservation/deleteFail.jsp");
+			request.getRequestDispatcher("/views/user/login.jsp").forward(request, response);
 		}
 	}
